+++ examples/rendering/borders.dart
@@ -6,6 +6,8 @@ import 'dart:sky' as sky;
 
 import 'package:sky/rendering/block.dart';
 import 'package:sky/rendering/box.dart';
+import 'package:sky/rendering/proxy_box.dart';
+import 'package:sky/rendering/shifted_box.dart';
 import 'package:sky/rendering/sky_binding.dart';
 
 void main() {
