+++ examples/rendering/render_paragraph.dart
@@ -5,11 +5,11 @@
 import 'dart:sky';
 
 import 'package:sky/painting/text_style.dart';
-import 'package:sky/rendering/box.dart';
 import 'package:sky/rendering/flex.dart';
 import 'package:sky/rendering/object.dart';
 import 'package:sky/rendering/paragraph.dart';
 import 'package:sky/rendering/sky_binding.dart';
+import 'package:sky/rendering/proxy_box.dart';
 
 import 'solid_color_box.dart';
