+++ examples/rendering/shadowed_box.dart
@@ -4,7 +4,8 @@
 
 import 'dart:sky';
 
-import 'package:sky/rendering/box.dart';
+import 'package:sky/rendering/proxy_box.dart';
+import 'package:sky/rendering/shifted_box.dart';
 import 'package:sky/rendering/sky_binding.dart';
 import 'package:sky/theme/colors.dart';
 import 'package:sky/theme/shadows.dart';
