+++ examples/rendering/solid_color_box.dart
@@ -3,8 +3,10 @@
 // found in the LICENSE file.
 
 import 'dart:sky' as sky;
-import 'package:sky/rendering/object.dart';
+
 import 'package:sky/rendering/box.dart';
+import 'package:sky/rendering/object.dart';
+import 'package:sky/rendering/proxy_box.dart';
 
 class RenderSolidColorBox extends RenderDecoratedBox {
   final Size desiredSize;
