+++ sky/packages/sky/lib/rendering/sky_binding.dart
@@ -4,10 +4,11 @@
 
 import 'dart:sky' as sky;
 
-import 'package:sky/base/scheduler.dart' as scheduler;
 import 'package:sky/base/hit_test.dart';
+import 'package:sky/base/scheduler.dart' as scheduler;
 import 'package:sky/rendering/box.dart';
 import 'package:sky/rendering/object.dart';
+import 'package:sky/rendering/view.dart';
 
 int _hammingWeight(int value) {
   if (value == 0)
