+++ sky/packages/sky/lib/widgets/scrollable.dart
@@ -10,6 +10,7 @@ import 'package:sky/animation/animated_simulation.dart';
 import 'package:sky/animation/animation_performance.dart';
 import 'package:sky/animation/scroll_behavior.dart';
 import 'package:sky/rendering/box.dart';
+import 'package:sky/rendering/viewport.dart';
 import 'package:sky/theme/view_configuration.dart' as config;
 import 'package:sky/widgets/basic.dart';
 import 'package:sky/widgets/block_viewport.dart';
