+++ sky/tests/layout/render_image.dart
@@ -1,7 +1,10 @@
 import 'dart:sky' as sky;
 
 import 'package:sky/rendering/box.dart';
+import 'package:sky/rendering/image.dart';
 import 'package:sky/rendering/object.dart';
+import 'package:sky/rendering/proxy_box.dart';
+import 'package:sky/rendering/shifted_box.dart';
 
 import '../resources/display_list.dart';
 import '../resources/third_party/unittest/unittest.dart';
