+++ sky/tests/raw/box_layout.dart
@@ -4,6 +4,8 @@ import 'dart:sky' as sky;
 import 'package:sky/rendering/box.dart';
 import 'package:sky/rendering/flex.dart';
 import 'package:sky/rendering/object.dart';
+import 'package:sky/rendering/proxy_box.dart';
+import 'package:sky/rendering/shifted_box.dart';
 
 import '../resources/display_list.dart';
