+++ sky/tests/raw/padding_deflate.dart
@@ -6,6 +6,8 @@ import 'dart:sky' as sky;
 
 import 'package:sky/rendering/box.dart';
 import 'package:sky/rendering/object.dart';
+import 'package:sky/rendering/proxy_box.dart';
+import 'package:sky/rendering/shifted_box.dart';
 
 import '../resources/display_list.dart';
 import '../resources/third_party/unittest/unittest.dart';
