+++ sky/tests/raw/render_box.dart
@@ -6,6 +6,7 @@ import 'dart:sky' as sky;
 
 import 'package:sky/rendering/box.dart';
 import 'package:sky/rendering/object.dart';
+import 'package:sky/rendering/proxy_box.dart';
 
 import '../resources/display_list.dart';
 import '../resources/third_party/unittest/unittest.dart';
